Create METAR_ESP32_07_12_22_Airservices_Melb_VNC.ino
This is the ESP32 METAR script written by John Pipe, with modifications to work with the Airservices Australia NAIPS API to pull regional Australia METARs (not just those available on SkyVector etc). The stations listed in this code are for those on the Melbourne VNC, in an order intended to be easiest to wire the LEDs around (see the PNG for a suggestion).